+++ metaseq/models/transformer_decoder.py
@@ -9,6 +9,8 @@ from typing import Any, Dict, List, Optional
 
 import torch
 import torch.nn as nn
+from torch import Tensor
+
 from metaseq import utils
 from metaseq.dataclass.constants import UNSPECIFIED_DOC_SEP
 from metaseq.distributed import utils as distributed_utils, fsdp_wrap
@@ -21,13 +23,12 @@ from metaseq.modules import (
     Linear,
 )
 from metaseq.modules.checkpoint_activations import checkpoint_wrapper
-from metaseq.modules.megatron import (
+from metaseq.modules.megatron.mpu import (
     LinearWithGradAccumulationAndAsyncCommunication,
     gather_from_tensor_model_parallel_region,
     scatter_to_sequence_parallel_region,
     copy_to_tensor_model_parallel_region,
 )
-from torch import Tensor
 
 logger = logging.getLogger(__name__)
