+++ metaseq/modules/megatron/mpu/layers.py
@@ -11,8 +11,8 @@ import torch.nn.functional as F
 import torch.nn.init as init
 from torch.nn.parameter import Parameter
 
-from metaseq.modules.megatron import (
-    get_global_memory_buffer,
+from metaseq.modules.megatron import get_global_memory_buffer
+from metaseq.modules.megatron.mpu import (
     get_tensor_model_parallel_group,
     get_tensor_model_parallel_rank,
     get_tensor_model_parallel_world_size,
