+++ metaseq/modules/megatron/mpu/mappings.py
@@ -10,8 +10,8 @@ import os
 
 import torch
 
-from metaseq.modules.megatron import (
-    get_global_memory_buffer,
+from metaseq.modules.megatron import get_global_memory_buffer
+from metaseq.modules.megatron.mpu import (
     get_tensor_model_parallel_group,
     get_tensor_model_parallel_world_size,
     get_tensor_model_parallel_rank,
