+++ metaseq/modules/megatron/mpu/random.py
@@ -12,7 +12,7 @@ import torch
 from torch import _C
 from torch.cuda import _lazy_call, device as device_ctx_manager
 
-from metaseq.modules.megatron import (
+from metaseq.modules.megatron.mpu import (
     get_data_parallel_rank,
     get_tensor_model_parallel_group,
     get_tensor_model_parallel_rank,
