+++ metaseq/modules/multihead_attention.py
@@ -8,20 +8,23 @@ from functools import partial
 from typing import Dict, Optional, Tuple
 
 import torch
+from torch import Tensor, nn
+
 from metaseq import utils
 from metaseq.dataclass.constants import AttentionVariants
 from metaseq.incremental_decoding_utils import with_incremental_state
 from metaseq.modules.dropout import Dropout
 from metaseq.modules.megatron import (
+    ScaledUpperTriangMaskedSoftmax,
+    ScaledMaskedSoftmax,
+)
+from metaseq.modules.megatron.mpu import (
     ColumnParallelLinear,
     RowParallelLinear,
     get_cuda_rng_tracker,
     get_tensor_model_parallel_world_size,
     split_tensor_along_last_dim,
-    ScaledUpperTriangMaskedSoftmax,
-    ScaledMaskedSoftmax,
 )
-from torch import Tensor, nn
 
 try:
     import xformers.ops as xops
